prettier

+++ contracts/external/solidly/IRouter.sol
@@ -37,13 +37,12 @@ interface IRouter {
     );
 
   function swapExactTokensForTokensSimple(
-    uint amountIn,
-    uint amountOutMin,
+    uint256 amountIn,
+    uint256 amountOutMin,
     address tokenFrom,
     address tokenTo,
     bool stable,
     address to,
-    uint deadline
-  ) external returns (uint[] memory amounts);
-
-  }
+    uint256 deadline
+  ) external returns (uint256[] memory amounts);
+}
